Set up routing for home and project page
