+++ package.json
@@ -11,8 +11,8 @@
     "watch": "tsc --watch",
     "fix": "tslint -p . --fix",
     "lint": "tslint -p .",
-    "format": "prettier --write 'src/{client,shared}/*.{ts,tsx,json,css}'",
-    "check-format": "prettier --check 'src/{client,shared}/*.{ts,tsx,json,css}'"
+    "format": "prettier --write 'src/{client,shared}/**/*.{ts,tsx,json,css}'",
+    "check-format": "prettier --check 'src/{client,shared}/**/*.{ts,tsx,json,css}'"
   },
   "dependencies": {
     "@types/jest": "24.0.22",
@@ -21,11 +21,15 @@
     "@types/react": "16.9.11",
     "@types/react-dom": "16.9.4",
     "@types/react-redux": "7.1.5",
+    "@types/react-router": "5.1.2",
+    "@types/react-router-dom": "5.1.2",
     "axios": "0.19.0",
     "mapbox-gl": "^1.6.1",
     "react": "16.11.0",
     "react-dom": "16.11.0",
     "react-redux": "7.1.3",
+    "react-router": "5.1.2",
+    "react-router-dom": "5.1.2",
     "react-scripts": "3.2.0",
     "redux": "4.0.4",
     "redux-loop": "5.0.1",
